Added View to HistoryView
